IUS-132: - Increased warning level and treat warning as error - fixed warnings

+++ library/src/iusPulseDict.c
@@ -217,7 +217,7 @@ iupd_t iusPulseDictLoad
     status = H5Gget_num_objs(grpid, &nobj);
 
     iupd_t dict = iusPulseDictCreate();
-    for (i = 0; i < nobj; i++)
+    for (i = 0; i < (int) nobj; i++)
     {
         H5Gget_objname_by_idx(grpid, (hsize_t) i,
                                     memb_name, (size_t) MAX_NAME);
